fixing splits model

+++ splits.go
@@ -2,8 +2,8 @@ package strava
 
 type Split struct {
 	Distance            float64 `json:"distance"`
-	ElapsedTime         int     `json:"elapsed_time"`
+	ElapsedTime         float64 `json:"elapsed_time"`
 	ElevationDifference float64 `json:"elevation_difference"`
-	MovingTime          int     `json:"moving_time"`
+	MovingTime          float64 `json:"moving_time"`
 	Split               int     `json:"split"`
 }
